src/bot_squared/requirements.txt
pyyaml
ts3
discord.py
mypy